Extract shared SdkDirectoryScope test helper
