+++ test/dotnet-aot.Tests/FormatForwardingAppTests.cs
@@ -2,30 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Microsoft.DotNet.Cli.Commands.Format;
-using Microsoft.DotNet.Cli.Utils;
 
 namespace Microsoft.DotNet.Cli.Tests;
 
 [TestClass]
 public class FormatForwardingAppTests
 {
-    private readonly struct SdkDirectoryScope : IDisposable
-    {
-        private readonly object? _previousSdkRoot = AppContext.GetData(SdkPaths.DataName);
-
-        public SdkDirectoryScope(string sdkDirectory)
-        {
-            AppContext.SetData(SdkPaths.DataName, sdkDirectory);
-            SdkPaths.ClearSdkDirectoryCacheForTests();
-        }
-
-        public void Dispose()
-        {
-            AppContext.SetData(SdkPaths.DataName, _previousSdkRoot);
-            SdkPaths.ClearSdkDirectoryCacheForTests();
-        }
-    }
-
     [TestMethod]
     public void FormatForwardingUsesVersionedSdkDirectory()
     {
