+++ test/dotnet-aot.Tests/MSBuildEvaluationTests.cs
@@ -15,23 +15,6 @@ namespace Microsoft.DotNet.Cli.Tests;
 [TestClass]
 public class MSBuildEvaluationTests
 {
-    private readonly struct SdkDirectoryScope : IDisposable
-    {
-        private readonly object? _previousSdkRoot = AppContext.GetData(SdkPaths.DataName);
-
-        public SdkDirectoryScope(string sdkDirectory)
-        {
-            AppContext.SetData(SdkPaths.DataName, sdkDirectory);
-            SdkPaths.ClearSdkDirectoryCacheForTests();
-        }
-
-        public void Dispose()
-        {
-            AppContext.SetData(SdkPaths.DataName, _previousSdkRoot);
-            SdkPaths.ClearSdkDirectoryCacheForTests();
-        }
-    }
-
     public TestContext TestContext { get; set; } = null!;
 
     [TestMethod]
